feat!: rename `default.json` to `marinatedconcrete.json` to support more configs
It does not make sense for us to have a default, but we'll have a few
different, simple configs to extend that have specific purposes.

This is a breaking change in that `default.json` has been renamed to
`marinatedconcrete.json`.  Consumers will need to update what they extend
for it continue to work.