fix: track versioned components in devcontainer spec
This fixes #269.  The original commit,
3143f97673cb8c899405c79c5fc6d2be279bd158, did not properly setup the
renovate file to parse these.  Since we now have a shared config that
can do that for us (pending release of 2.0.0 in #270), this change
undoes the work in the original commit and instead depends on our shared
config to pick it up.

Validated with these two tests of our updated config:
* https://regex101.com/r/SE5Vfw/1
* https://regex101.com/r/ocT8Ie/1