!feat: generate kube-vip manifest (#329)
This is a breaking change because it potentially breaks patching, as the
namespace may have to be added to patch files as seen in the test
updates in this change.

This uses kube-vip's manifest generation utility, which their docs
really push you to use. This is versioned with the kube-vip release in
our repository, which is synced with renovate.

This change also documents the required and optional patches for
kube-vip.

Tested renovate config settings with https://regex101.com/r/FwnumZ/1