feat: Ensure each thing we release is not grouped by renovate (#481)
Renovate groups all the updates from this repository into a single Pull
Request. Ultimately, that's because they all share the same branch name.
We can override that by either changing the branch name (which can break
other packages, so it is a bad idea), or by specifying a group for each
package.

This solution is verbose, but it works great. I was able to pull out a
Unifi update in https://github.com/sdwilsh/ansible-playbooks/pull/1191,
as an example.